Working Command Line Version

+++ src/expression/model/expression/Exponential.java
@@ -4,9 +4,9 @@ import expression.Expression;
 
 public class Exponential extends Expression {
     public Expression left;
-    public Float operator;
+    public Double operator;
 
-    public Exponential(Expression left, Float operator) {
+    public Exponential(Expression left, Double operator) {
         this.left = left;
         this.operator = operator;
     }
